+++ client/src/App.js
@@ -1,6 +1,8 @@
 import React from "react";
 import SignUp from "./components/Sing-up";
-import Cart from "./components/Cart"
+import Cart from "./components/Cart";
+
+
 function App() {
   return (
     <div>
